ai/pyproject.toml
[tool.poetry]
name = "ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = "^0.0.51"}
kinmi-speak = {path = "packages/kinmi-speak", develop = true}
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
langchain-openai = "^0.1.1"
langchain-community = "^0.0.31"
langchain-anthropic = "^0.1.6"
langmem = "^0.0.3"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
